# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeTextBody {
  text: string;
  /** @nullable */
  sourceUrl?: string | null;
}

export interface AnalyzeUrlBody {
  url: string;
}

export type AnalysisResultPrediction =
  (typeof AnalysisResultPrediction)[keyof typeof AnalysisResultPrediction];

export const AnalysisResultPrediction = {
  Fake: "Fake",
  Real: "Real",
} as const;

export type SourceCredibilityLabel =
  (typeof SourceCredibilityLabel)[keyof typeof SourceCredibilityLabel];

export const SourceCredibilityLabel = {
  Trusted: "Trusted",
  Untrusted: "Untrusted",
  Unknown: "Unknown",
} as const;

export interface SourceCredibility {
  /** @nullable */
  domain: string | null;
  score: number;
  label: SourceCredibilityLabel;
}

export interface AnalysisResult {
  id: number;
  prediction: AnalysisResultPrediction;
  confidence: number;
  reasons: string[];
  suspiciousWords: string[];
  sourceCredibility: SourceCredibility;
  inputText: string;
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  factualContext?: string | null;
  analyzedAt: string;
}

export interface AnalysisRecord {
  id: number;
  prediction: string;
  confidence: number;
  reasons: string[];
  suspiciousWords: string[];
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  factualContext?: string | null;
  /** @nullable */
  sourceDomain?: string | null;
  sourceCredibilityScore: number;
  sourceCredibilityLabel: string;
  inputText: string;
  /** @nullable */
  feedbackCorrect?: boolean | null;
  createdAt: string;
}

export interface HistoryResponse {
  items: AnalysisRecord[];
  total: number;
}

export interface FeedbackBody {
  analysisId: number;
  correct: boolean;
}

export interface FeedbackRecord {
  id: number;
  analysisId: number;
  correct: boolean;
  createdAt: string;
}

export interface StatsResponse {
  totalAnalyzed: number;
  fakeCount: number;
  realCount: number;
  avgConfidence: number;
  /** @nullable */
  accuracyFromFeedback?: number | null;
  recentActivity: AnalysisRecord[];
}

export type GetHistoryParams = {
  limit?: number;
  offset?: number;
};