# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Analyze news text
 */
export const AnalyzeTextBody = zod.object({
  text: zod.string(),
  sourceUrl: zod.string().nullish(),
});

export const AnalyzeTextResponse = zod.object({
  id: zod.number(),
  prediction: zod.enum(["Fake", "Real"]),
  confidence: zod.number(),
  reasons: zod.array(zod.string()),
  suspiciousWords: zod.array(zod.string()),
  sourceCredibility: zod.object({
    domain: zod.string().nullable(),
    score: zod.number(),
    label: zod.enum(["Trusted", "Untrusted", "Unknown"]),
  }),
  inputText: zod.string(),
  sourceUrl: zod.string().nullish(),
  analyzedAt: zod.string(),
});

/**
 * @summary Analyze news article from URL
 */
export const AnalyzeUrlBody = zod.object({
  url: zod.string(),
});

export const AnalyzeUrlResponse = zod.object({
  id: zod.number(),
  prediction: zod.enum(["Fake", "Real"]),
  confidence: zod.number(),
  reasons: zod.array(zod.string()),
  suspiciousWords: zod.array(zod.string()),
  sourceCredibility: zod.object({
    domain: zod.string().nullable(),
    score: zod.number(),
    label: zod.enum(["Trusted", "Untrusted", "Unknown"]),
  }),
  inputText: zod.string(),
  sourceUrl: zod.string().nullish(),
  analyzedAt: zod.string(),
});

/**
 * @summary Get analysis history
 */
export const getHistoryQueryLimitDefault = 20;
export const getHistoryQueryOffsetDefault = 0;

export const GetHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getHistoryQueryLimitDefault),
  offset: zod.coerce.number().default(getHistoryQueryOffsetDefault),
});

export const GetHistoryResponse = zod.object({
  items: zod.array(
    zod.object({
      id: zod.number(),
      prediction: zod.string(),
      confidence: zod.number(),
      reasons: zod.array(zod.string()),
      suspiciousWords: zod.array(zod.string()),
      sourceUrl: zod.string().nullish(),
      sourceDomain: zod.string().nullish(),
      sourceCredibilityScore: zod.number(),
      sourceCredibilityLabel: zod.string(),
      inputText: zod.string(),
      feedbackCorrect: zod.boolean().nullish(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get a single analysis by ID
 */
export const GetHistoryItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetHistoryItemResponse = zod.object({
  id: zod.number(),
  prediction: zod.string(),
  confidence: zod.number(),
  reasons: zod.array(zod.string()),
  suspiciousWords: zod.array(zod.string()),
  sourceUrl: zod.string().nullish(),
  sourceDomain: zod.string().nullish(),
  sourceCredibilityScore: zod.number(),
  sourceCredibilityLabel: zod.string(),
  inputText: zod.string(),
  feedbackCorrect: zod.boolean().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Submit feedback on an analysis
 */
export const SubmitFeedbackBody = zod.object({
  analysisId: zod.number(),
  correct: zod.boolean(),
});

/**
 * @summary Get overall stats dashboard summary
 */
export const GetStatsResponse = zod.object({
  totalAnalyzed: zod.number(),
  fakeCount: zod.number(),
  realCount: zod.number(),
  avgConfidence: zod.number(),
  accuracyFromFeedback: zod.number().nullish(),
  recentActivity: zod.array(
    zod.object({
      id: zod.number(),
      prediction: zod.string(),
      confidence: zod.number(),
      reasons: zod.array(zod.string()),
      suspiciousWords: zod.array(zod.string()),
      sourceUrl: zod.string().nullish(),
      sourceDomain: zod.string().nullish(),
      sourceCredibilityScore: zod.number(),
      sourceCredibilityLabel: zod.string(),
      inputText: zod.string(),
      feedbackCorrect: zod.boolean().nullish(),
      createdAt: zod.string(),
    }),
  ),
});