# lib/api-zod/src/generated/types/analysisRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRecord {
  id: number;
  prediction: string;
  confidence: number;
  reasons: string[];
  suspiciousWords: string[];
  /** @nullable */
  sourceUrl?: string | null;
  /** @nullable */
  sourceDomain?: string | null;
  sourceCredibilityScore: number;
  sourceCredibilityLabel: string;
  inputText: string;
  /** @nullable */
  feedbackCorrect?: boolean | null;
  createdAt: string;
}