# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultPrediction } from "./analysisResultPrediction";
import type { SourceCredibility } from "./sourceCredibility";

export interface AnalysisResult {
  id: number;
  prediction: AnalysisResultPrediction;
  confidence: number;
  reasons: string[];
  suspiciousWords: string[];
  sourceCredibility: SourceCredibility;
  inputText: string;
  /** @nullable */
  sourceUrl?: string | null;
  analyzedAt: string;
}