# lib/api-zod/src/generated/types/feedbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackBody {
  analysisId: number;
  correct: boolean;
}