# lib/api-zod/src/generated/types/feedbackRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackRecord {
  id: number;
  analysisId: number;
  correct: boolean;
  createdAt: string;
}