# lib/api-zod/src/generated/types/historyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRecord } from "./analysisRecord";

export interface HistoryResponse {
  items: AnalysisRecord[];
  total: number;
}