# lib/api-zod/src/generated/types/sourceCredibility.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
import type { SourceCredibilityLabel } from "./sourceCredibilityLabel";

export interface SourceCredibility {
  /** @nullable */
  domain: string | null;
  score: number;
  label: SourceCredibilityLabel;
}