# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fake News Detector API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisRecord } from "./analysisRecord";

export interface StatsResponse {
  totalAnalyzed: number;
  fakeCount: number;
  realCount: number;
  avgConfidence: number;
  /** @nullable */
  accuracyFromFeedback?: number | null;
  recentActivity: AnalysisRecord[];
}